refactor: LottoNumbersGenerator 클래스와 그 테스트 코드 패키지 변경
domain에서 util로 패키지 변경
