+++ src/test/java/lotto/util/LottoNumbersGeneratorTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.util;
 
 import static org.assertj.core.api.Assertions.assertThat;
